XRDDEV-1949 Rename router file so it's easier to find
