+++ src/centralserver/admin-ui/src/router/router.ts
